refactor:  abstract the debounced search of shops into DebouncedShopSearchMixin
