+++ lib/widgets/tabs/shop_card_tab.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
+import '../../mixins/debounced_shop_search_mixin.dart';
 import '../../providers/shop_search_name_provider.dart';
 import '../../repositories/shop_repository.dart';
-import '../../utils/debouncer.dart';
 import '../loading/lazy_list_view.dart';
 import '../loading/shop_card_skeleton.dart';
 import '../shop_card.dart';
@@ -17,32 +17,13 @@ class ShopCardTab extends ConsumerStatefulWidget {
   ConsumerState<ShopCardTab> createState() => _ShopCardTabState();
 }
 
-class _ShopCardTabState extends ConsumerState<ShopCardTab> {
-  List<Shop> shops = [];
-  bool isLoading = false;
-
-  late final debouncedGetShops = Debouncer(milliseconds: 500, action: getShops);
-
-  Future<void> getShops() async {
-    setState(() => isLoading = true);
-
-    final searchName = ref.read(shopSearchNameProvider);
-
-    try {
-      shops = await ref.read(shopRepositoryProvider).getShops(
-            type: widget.shopType,
-            name: searchName,
-          );
-    } finally {
-      setState(() => isLoading = false);
-    }
-  }
+class _ShopCardTabState extends ConsumerState<ShopCardTab>
+    with DebouncedShopSearchMixin {
+  @override
+  String? get shopName => ref.read(shopSearchNameProvider);
 
   @override
-  void initState() {
-    super.initState();
-    getShops();
-  }
+  ShopType get shopType => widget.shopType;
 
   @override
   Widget build(BuildContext context) {
